+++ src/utils/xrCompress/xrCompress.cpp
@@ -151,10 +151,10 @@ xrCompressor::ALIAS* xrCompressor::testALIAS(IReader* base, u32 crc, u32& a_test
 void xrCompressor::write_file_header(
     LPCSTR file_name, const u32& crc, const u32& ptr, const u32& size_real, const u32& size_compressed)
 {
-    u32 file_name_size = (xr_strlen(file_name) + 0) * sizeof(char);
-    u32 buffer_size = file_name_size + 4 * sizeof(u32);
+    size_t file_name_size = (xr_strlen(file_name) + 0) * sizeof(char);
+    size_t buffer_size = file_name_size + 4 * sizeof(u32);
     VERIFY(buffer_size <= 65535);
-    u32 full_buffer_size = buffer_size + sizeof(u16);
+    size_t full_buffer_size = buffer_size + sizeof(u16);
     u8* buffer = (u8*)_alloca(full_buffer_size);
     u8* buffer_start = buffer;
     *(u16*)buffer = (u16)buffer_size;
@@ -436,7 +436,7 @@ void xrCompressor::PerformWork()
         if (!bStoreFiles)
             c_heap = xr_alloc<u8>(LZO1X_999_MEM_COMPRESS);
 
-        for (u32 it = 0; it < files_list->size(); it++)
+        for (size_t it = 0; it < files_list->size(); it++)
         {
             xr_sprintf(caption, "Compress files: %d/%d - %d%%", it, files_list->size(), (it * 100) / files_list->size());
             SetWindowText(GetConsoleWindow(), caption);
@@ -543,7 +543,7 @@ void xrCompressor::ProcessLTX(CInifile& ltx)
             string_path path;
             LPCSTR _path = 0 == xr_strcmp(it.first.c_str(), ".\\") ? "" : it.first.c_str();
             xr_strcpy(path, _path);
-            u32 path_len = xr_strlen(path);
+            size_t path_len = xr_strlen(path);
             if ((0 != path_len) && (path[path_len - 1] != '\\'))
                 xr_strcat(path, "\\");
